+++ sdk/issuingBillingInvoice/index.js
@@ -1,0 +1,6 @@
+const issuingBillingInvoice = require('./issuingBillingInvoice.js');
+
+exports.get = issuingBillingInvoice.get;
+exports.query = issuingBillingInvoice.query;
+exports.page = issuingBillingInvoice.page;
+exports.IssuingBillingInvoice = issuingBillingInvoice.IssuingBillingInvoice;