+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - installmentCount attribute to IssuingPurchase resource
+- IssuingBillingInvoice resource
+- IssuingBillingTransaction resource
+- activationCode and url attributes to IssuingToken resource
 
 ## [0.18.0] - 2026-05-04
 ### Added